python/requirements.txt
# Core dependencies
instructor>=0.4.0
openai>=1.12.0
pydantic>=2.6.0
pydantic-ai>=0.1.0
anthropic>=0.18.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
rich>=13.7.0
tqdm>=4.66.0

# CLI interface
rich==13.7.0

# Interface with C++ code
pybind11>=2.10.0

# Database interfaces
duckdb>=0.9.0
qdrant-client>=1.13.3

# Utilities
numpy>=1.26.0
pandas>=2.2.0
python-multipart>=0.0.9
uvicorn[standard]>=0.27.0
fastapi>=0.109.0

# Visualization & Analysis
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
scipy>=1.12.0
scikit-learn>=1.4.0
pillow>=10.2.0

# New additions
pydantic>=2.0.0
pydantic-ai>=0.0.46
fastapi>=0.98.0
uvicorn>=0.22.0
python-multipart>=0.0.6
duckdb>=0.10.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.16.0
rich>=13.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
instructor>=1.0.0 